+++ docs/md/releasenotes.md
@@ -5,6 +5,7 @@
 - `KotlinTypeParameterVistor#visitClassParameter(Clazz, KotlinMetadata, KotlinTypeParameterMetadata)` now has the correct signature: `KotlinTypeParameterVistor#visitClassParameter(Clazz, KotlinClassKindMetadata, KotlinTypeParameterMetadata)`. 
 - Rename `AllKotlinPropertiesVisitor` to `AllPropertyVisitor`.
 - Rename `AllConstructorsVisitor` to `AllConstructorVisitor`.
+- Rename `AllFunctionsVisitor` to `AllFunctionVisitor`.
 
 ## Version 7.1.1